Major bugfixes regarding parsing process and minor GUI fixes.
Now leo-lookup is able to look up 2 or more words, like "to go".
In addition, the first result will now be shown and not throwed out.
Additionally, the user can see the amount of found results in the main window.
For debugging reasons, the logging facility has been integrated and messages
will be written in leo-lookup.log in the current directory.