Added major changes, like plugins and plans regarding plugins and their interface.
Added plugin for using with jMemorize 1.3.0 (http://jmemorize.org/), but
at this commit, it cannot be used with leo-lookup, because the support of
plugins is not yet implemented in leo-lookup.py. In addition, the directory
structure has changed. There are now two directories: one for planning
documents (like diagrams or text documents) and one for plugins.